Cast scale factor in pattern.scale; add unit tests
Instead of checking whether the 'by' argument is an array or a float,
simply recast it as an array. If the input is a float, the positions are
scaled by a float. If the input is an ndarray with shape=(3,) then you
get different scaling for x, y, and z as intended.

Also add a few unit tests:
    - Pass length 2 and length 4 arrays; assert failure
    - Pass length 2 and length 4 arrays and try to scale a pattern with
      2 and 4 points; assert failure
    - Try to scale a 100-point pattern by a 100 dimensional vector;
      assert faliure