Add more descriptive error when _clone_bonds fails
Cloning of a Compound will fail if it contains bonds to Particles
that reside outside of the Compound's containment hierarchy (as
these Particles are not included in the clone). Rather than simply
raising a KeyError when the bond cloning fails, a more descriptive
error message is now provided.

A unit test has been added to check for this error message.