Intelligent default for area per lipid
Multicomponent systems were having a difficult time equilibrating with the given APL default. This was resolved by creating an APL window that is dependent on the concentration of phospholipids in the bilayer, so multicomponent systems with a higher concentration of smaller molecules (alc, ffa) are given a window of smaller APL values. Randomization maintained.