Unit test addition and small fix to Compound
Compound.center had an issue using numpy.any()

0 evaluates to false in python and if the only coordinate
was [0,0,0] in the compound, the center was set to None, not
[0,0,0]. Instead of using numpy.any(), the length of the xyz array was
checked. If the array size is nonzero, the center is calculated.

More unit tests to ensure that the compounds are populated correctly.

Small change in lattice code from using a particle to compound.