Lattice overhaul (#386)
* Lattice changes and clean up

Changed data structures for many of the underlying lattice attributes.

bugfix when populating a lattice with non-cubic geometries.

* Overhaul and fix lattice point transformation

Lattice attributes cleaned up and mostly moved to Numpy
arrays.

Simpler inputs.

Fixed change of basis error when populating and extending the base lattice
object.

* Added angles to lattice no matter if vectors or angles passed in.

Previously, if the user developed the Lattice object through definition of
cell vectors, the bravais angles corresponding to those vectors would not be saved.

The inverse was true when the user passed in the bravais angles
  * alpha, beta, gamma

Fixed many unit tests for the Lattice class after the new revisions

* Additional Test

* More unit tests and small change to angle setting

Added more unit tests for the populate function.

Ensure that the angles are set properly when deriving lattice
from lattice vectors.

* Unit test addition and small fix to Compound

Compound.center had an issue using numpy.any()

0 evaluates to false in python and if the only coordinate
was [0,0,0] in the compound, the center was set to None, not
[0,0,0]. Instead of using numpy.any(), the length of the xyz array was
checked. If the array size is nonzero, the center is calculated.

More unit tests to ensure that the compounds are populated correctly.

Small change in lattice code from using a particle to compound.

* Changed the examples
Examples changed to reflect the new lattice structures

* Added __init__ file to tests directory

For some reason, pytest could not figure out the proper
directory to import the tests from. The 2.7 builds would fail without
this empty file being added.

Unsure what this does, but it seems to work on my 2.7 environment.

* Corrected improper basis positions

* Tutorial notebook and setting periodicity

* Minor cleanup

* Array size fixes

* User warning for periodicity, test for setting periodicity

* Final cleanup

* More coverage

* Coverage, and parameter docstring tweaking