Fixed bug in alignment of subports in __init__
Ports along antiparallel vectors should have 'up' and 'down' subports
at equivalent positions, except for reverse labels (i.e. the 'up'
subport of a port oriented in one direction will have the same xyz
positions as the 'down' port oriented in the opposite direction).
However, the previous routine to initialize the subports performed
this incorrectly. The committed changes fix this functionality. The
alignment issues appeared to occur if the z-coordinate of the
orientation vector was less than zero. I could not figure out the
proper rotations to fix this, so as a workaround, in cases where the
z-coordinate of the orientation vector is less than zero, the subports
are created as if the orientation vector was in the opoosite direction
and the labels are switched.

Additional changes include updating the docstring of the Port __init__,
changing the default `orientation` from `None` to `[0, 1, 0]` (since
this was done internally anyhow), rotating the port 180 degrees around
z instead of x if `[0, -1, 0]` is provided as the orientation, and
replacing `np.array_equal` to `np.allclose` when checking for specific
orientation vectors.

Highlights
* `np.array_equal` replaced with `np.allclose` in orientation checks
* Proper handling of subport positioning when orientation vectors with
  a -z coordinate are passed
* Updated docstring of Port __init__
* Default `orientation` changed from `None` to `[0, 1, 0]`
* Rotating Port 180 degrees around z instead of x if `[0, -1, 0]` is
  provided as the orientation.