Added tests for reconnecting Compounds
Tests have been added that ensure that when a bond is broken
between two Compounds and the Compounds are reconnected, that the
relative orientation of the Compounds is unchanged. This is tested
by connecting two CHF molecules and checking that the FCCF dihedral
remains the same after disconnecting the reconnecting the molecules.
These tests cover the edge cases of connections along the three
coordinate axes as well as along five random vectors.