Provide segfault trace for pytest
Currently all python 3.5 osx builds on Travis are
failing at the collection stage of pytest due to
a SIGABRT signal being raised.

We currently do not have a way to interpret where this
is occuring since we cannot replicate on any osx 3.5
machines in the office.

There is a pytest plugin called `pytest-faulthandler` that
enables the `faulthandler` module for pytest.

This should provide a more verbose output when the system recieves
a signal like this in the future.