lattice populated as coordinates are determined in real time/
The populate routine has been changed so that as coordinates of particles
are determined particles are added to the Compound lattice, thus the
"cell" is no longer created and iterated through. I expect to see speed
improvements. The next step is to functionalize terminal groups as the
corresponding particles are added.