Enforce captialization standards in to_parmed()
Enforce uppercase,lowercase formatting for element abbreviations
(Si, not SI, Cl not CL).

When converting a `Compound` to a `ParmEd` structure using the
`Compound.to_parmed()` method, there is a lookup section to
infer the elements present in the compound. If the element names are
fully capitalized (which can happen when reading in from certain mol2
files, and others), the inference does not correctly determine the
chemistry. In the above examples SI would be inferred as Sulfur (S), and
CL would be inferred as Carbon (C).

This can induce issues when trying to go from an mBuild Compound, to
a ParmEd structure, to an atomtyped system using Foyer.

This is not a catch-all fix, as non-elemental names that are not
prefaced with an underscore (which they should be, but sometimes this is
missed when making a compound) will undergo this same treatment.