Include support for the GSD 2.0 file io syntax
With the update to GSD 2.X, the way we have been writing out the GSD
snapshot is no longer available.

The syntax has been updated to the non-deprecated and recommended way to
write out to GSD files. This feature was added in ver 1.2 of GSD, which
is why the version in requirements-dev.txt has been pinned to >=1.2