provide basic framework for loading in cif files
Currently, mBuild has no way to read in CIF files, which are very useful
for many structured materials with defined repeat units.

This provides an inital attempt to include this functionality using the
`pycifrw` library. Which is maintained by the
Australian Nuclear Science and Technology Organisation (ANSTO).This
seems to be a pretty well tested CIF library, with minimal dependencies.

Note that this commit does not include all the additional features
necessary to successfully generate structures with the proper repeat
units, since the symmetry operations have not been included.

This current commit can load in CIF files, determine the Bravais angles,
the lattice spacings, and the coordinates of the listed atom sites. It
should be noted that this information does not return a full crystal
structure, since each site should also have symmetry operations denoted
by their Hermann-Mauguin notation. This is provided usually in each CIF
file, but there are a lot of complications to support this fully.