Change internal reader to the garnett package
The cif reader previously had no support for symmetry operations and
its' element detection was very limited.

This commit replaces the backend for loading cif files and uses the
garnett package. However, this is not the complete story, the user will
still need pycifrw to load in the cif files as that is the library
garnett's cif reader also uses.

The element detection here exists, but if garnett can not find elemental
names, the cooridate positions are replaced with `A` as their element
name. For a lot of systems that have ions in them, the cif file
usually adds additional information like `Si4+` for a certain silicon
location. This is not parsed, but replaced with `A`. The user will need
to verify their cif files to ensure that non-elemental names are
handled properly.