WIP - add Minor Changes to Dockerfile and Docker Documentation
1. Previously, Dockerfile was installing python twice, this has been
changed to consolidate install step with environment creation. The
Dockerfile also create a directory called mbuild-notebooks as a default
directory to start jupyter-notebooks from the container. Additionally,
the CMD for docker image is changed to switch user inorder to be able
to mount directory with proper permissions from a custom entrypoint.

2. A sepearte section on mbuild docs is added as a docker tutorial