Update unit tests and add coordinate wrapping
Previously the unit tests for this feature was lacking. However,
switching to `garnett` for the cif reading simplifies the loading
process and preprocessing of cif files on `mbuild`'s side.

When loading cif files, a wrap_coords argument is now available for
load_cif. If the fractional cif coordinates are ~1 additional unit
lattice (-1 to 0, and 1 to 2) outside the unit cell, these will be
wrapped back into the first unit lattice.

This feature is turned off by default.