provide basic framework for loading in cif files (#689)
* provide basic framework for loading in cif files

Currently, mBuild has no way to read in CIF files, which are very useful
for many structured materials with defined repeat units.

This provides an inital attempt to include this functionality using the
`pycifrw` library. Which is maintained by the
Australian Nuclear Science and Technology Organisation (ANSTO).This
seems to be a pretty well tested CIF library, with minimal dependencies.

Note that this commit does not include all the additional features
necessary to successfully generate structures with the proper repeat
units, since the symmetry operations have not been included.

This current commit can load in CIF files, determine the Bravais angles,
the lattice spacings, and the coordinates of the listed atom sites. It
should be noted that this information does not return a full crystal
structure, since each site should also have symmetry operations denoted
by their Hermann-Mauguin notation. This is provided usually in each CIF
file, but there are a lot of complications to support this fully.

* Change internal reader to the garnett package

The cif reader previously had no support for symmetry operations and
its' element detection was very limited.

This commit replaces the backend for loading cif files and uses the
garnett package. However, this is not the complete story, the user will
still need pycifrw to load in the cif files as that is the library
garnett's cif reader also uses.

The element detection here exists, but if garnett can not find elemental
names, the cooridate positions are replaced with `A` as their element
name. For a lot of systems that have ions in them, the cif file
usually adds additional information like `Si4+` for a certain silicon
location. This is not parsed, but replaced with `A`. The user will need
to verify their cif files to ensure that non-elemental names are
handled properly.

* Update tests and importer

* Update mbuild/utils/io.py

Co-Authored-By: Bradley Dice <bdice@bradleydice.com>

* Update mbuild/lattice.py

Co-Authored-By: Bradley Dice <bdice@bradleydice.com>

* Update requirements-dev.txt

Co-Authored-By: Bradley Dice <bdice@bradleydice.com>

* Retrieve correct box dimensions and convert to nm

* Update unit tests and add coordinate wrapping

Previously the unit tests for this feature was lacking. However,
switching to `garnett` for the cif reading simplifies the loading
process and preprocessing of cif files on `mbuild`'s side.

When loading cif files, a wrap_coords argument is now available for
load_cif. If the fractional cif coordinates are ~1 additional unit
lattice (-1 to 0, and 1 to 2) outside the unit cell, these will be
wrapped back into the first unit lattice.

This feature is turned off by default.

* test for pycifrw and improve warnings
The requirements-dev.txt file included garnett, however, pycifrw (the
cif reading backend) is a soft dependency.

Pycifrw is now required along with garnett for cif reading capabilites.
A `has_pycifrw` test has also been added.

As well as a minor update to the lattice class when warning the user if
the lattice is not rectangular (alpha, beta, gamma) == 90.

* Conditionally import garnett and pycifrw for cif

Previously, there was not a check in the cif file reader for garnett and
its soft dependency pycifrw.

This has now been added, as well as instructions to install both of
these packages through conda.

Co-authored-by: Bradley Dice <bdice@bradleydice.com>