Include missing transpose of cif box_matrix
Currently, when creating a `Lattice` from a CIF file, the information
that is used to construct the lattice is the `box_matrix` from the
Garnett CIF reader.

The box_matrix has the a,b,c vectors of the lattice arranged in a
columnar format:

| ax | bx | cx |
| --- | --- | --- |
| ay | by |cy|
| az | bz |cz|

The a vector is the first column, b the next, etc.

mBuild constructs its Lattice's with a box matrix in row-ordered
formatting, which is the transpose of the above matrix.

The transpose was never completed before constructing a lattice, which
led to numerous issues regarding the lattice dimensions.

This fix transposes the matrix before being passed to the `Lattice`
constructor.