Introduce beginning overhaul of the Box object
Previously, the Box object was quite lacking in features.

Primarily the Box could only be expected to reasonably support
orthogonal boxes. Although there was support introduced to include
angles to deform the box into a desired parallelepiped, this has not been
implemented.

This is the beginning steps to overhaul the Box class to make triclinic
systems a first-class citizen in mBuild.

Note the use of classmethods to define alternative constructors for the
Box. There are many ways to define a Box in the scientific simulation
world, and as many of these variations we are trying to support. As long
as these values can be coerced into a set of box vectors, additional
constructors can be added as well.