Update classmethods with box initialization logic
During this overhaul of the Box class, multiple ways to construct an
instance of a Box are being supported with the use of different
classmethods.

During the initial phase of this development, many of the ways to
construct this object were defined but left as methods that do nothing.
This commit updates the constructors when given lo and hi x,y,z values
as well as the tilt factors of the box. This is a common way to
describe boxes in LAMMPS.

Also, the constructor when given minimum x,y,z values, maximum x,y,z
values, and the angles defining the deformation of the orthogonal box
into a parallelepiped.

Note, these do not have unit tests yet, but a majority of the logic
draws from the `Lattice` class in mBuild, which has similar methods, as
well as from various simulation engines like LAMMPS and HOOMD-Blue which
extensively document various box transformations