Include logic to validate box_vectors
Previously, the method to validate the box vectors was incomplete, this
has now been changed.

The full logic to validate the box vectors according to the standards
commonly implemented in the simulation community have been put in place.
If there are just algebraic transformations to fix the inputted vectors,
these are done automatically, but will warn the user. If it cannot be
fixed, an MBuildError is raised.

Note to properly define a parallelepiped with 3 vectors requires:
1. It is a 3x3 matrix that can be coerced into a numpy array of floats
2. Vectors are in a right-handed basis (determinant of matrix is +)
3. The first vector is aligned along the [1,0,0] direction
4. The second vector in aligned along the xy plane
5. The third vector can align freely in the x,y, and +z direction