Include BSD 3-clause from garnett
To normalize the box vectors, garnett, a file reader and writer from the
GLotzer group, had a very clean method to orient the box vectors in a
right handed method as well as orienting the first vector along (1,0,0),
the second along (1,1,0), and the third along (1,1,1).

Parts of this code could be used to re-orient the Box vectors into the
proper orientation as defined above. Following the copyright rules of
the BSD 3-clause license garnett is released under, a copy of garnett's
BSD license notice as well as a more detailed description of what
exactly was copied from garnett's source code and implemented in Box.py
is provided.