Include tests for left handed boxes and getters
Tests that check when a left handed matrix is provided by the user have
been implemented.

This checks that the correct UserWarning as message is raised, and the
messsage itself is validated.

More getters and the box array of Lx,Ly,Lz,xy,xz,yz are attributes of
the Box now.