Include getters for the box parameters
If you wanted the parameters of the Box (Lx,Ly,Lz,xy,xz,yz), the only
way was to return all of these at once using Box.box_parameters().

This is less ideal when trying to check a single value like Lz, for
example.

Getters for each box parameter are now available.