Re-factor readers and writers for Compound
With the box overhaul, a lot of the current implementations of the
`Box` object are not properly formatted.

Initial attempts to refactor these methods have begun, although this
commit is not fully functional yet. Quite a few assumptions are being
made about the box dimensions and centers (we previously used `mins` and
`maxs` of a box to define the upper and lower bounds of a system.

This is a critical difference that needs to be addressed: Should a `Box`
contain information like mins and max locations in space for a
`Compound` when writing out to a simulation format file?

This seems more like a method or attribute of a Compound, not a Box.