Add property accessors to Box, update packing
A majority of the failing unit tests were less with the Box logic
itself, but with the pytest fixtures that relied on packing.py (which
had not yet been updated).

Packing.py has been updated, as well as a majority of the tests as well.

There is one failing test that is related to the aspect_ratio flag for
fill_box in packing.py