Support json serialization of elements
With the addition of the mb.Compound.element attribute, the various supported formats for reading/writing to mb.Compounds needed to be updated.

This commit includes support for elements (ele.element or a string
representation) when being serialized or de-serialized to/from the JSON
format.
The relevant unit tests were also updated.