Use correct lattice spacing when reading from cif
Previously, when loading in from a cif file, the Lattice object used the
Box.Lx, Box.Ly, Box.Lz from the garnett trajectory; this value was in
eucledian space, not lattice space (which is what the lattice object
operates in).

This oversight would lead to discrepancies when loading in a cif file
versus building the same lattice object manually. Now, the lattice-space
lattice spacings are correctly set when loading a cif file.
This matches current manual lattice behavior.

An additional test is now included to ensure that whether building
manually, or from a cif file, that the lattice objects are functionally
equivalent.