Migrate existing SMILES loader to RDKit
Currently, our mbuild.load method can generate Compounds from a SMILES
string using openbabel as the backend loader, this is not great for
reproducible starting configurations as there is no current way to
control the PRNG seed.

Without being able to control the PRNG seed, the xyz coords are always
different between subsequent mb.load calls on a single SMILES string.
RDKit, however, does allow the user to control the PRNG seed when using
its SMILES generation methods. This PR is an initial attempt to replace
openbabel as our de facto SMILES loading backend with RDKit, and allow
the users to control the random seed to further increase the
reproducibility of system initialization.