Fix periodic_dist bug, update default box init
There was an issue with the min_periodic_distance calculation related to
the updates to the Box object and the deprecation of the `periodicity`
calculations.

This error has been fixed, some of the code base has been formatted
using `black`. The default `box` constructor has been switched from
providing a box matrix, to lengths and angles.