Feat/lammpswriter improvements (#924)
* Incorporate quality of life updates for lammps

Our lammpswriter is currently missing some minor quality of life fixes
as well as some slightly unexpected behavior for approximately
orthorhombic systems.

This PR allows the user to define a custom label for the `pair_coeffs`
section, which is useful if using a different pair style that is not
just the `lj` standard lammps fix.

The atoms section was missing a label as well defining the atom style,
this is commonly used by various visualization tools to better inform
the tools what type of data to expect.

The triclinic box vs orthorhomic box check had too high of a tolerance.
This has been updated. This is important since triclinic boxes in lammps
are less performant compared to rectangular boxes.

When writing out the masses of the atoms in the system, we do not use
the atom_type mass, but the elemental mass. A check is applied to use
the atom_type mass if possible, if not, elemental mass.

* Include unit tests for new lammps features

Unit tests for these previously discussed quality of life fixes have
been added. As well as a small fix for the lj unit style.

* Add warnings and fallback to use atomic mass if atomtype mass cannot be determined.