Update PeriodicKDTree to support box overhaul
There were certain situations where the periodic KDTree was not updated
properly to support the updates to the periodicity and the box overhaul.

There is still one test case failing for `TiledCompound`, but this PR
helps a lot with generating bonds. Once the `TiledCompound` error is
fixed, this should be good to merge.