Include moleculeID offset in lammpswriter
LAMMPS handles moleculeID's as an additional set of data to link atoms
with specific molecules (similar to residues in other engines).

However, there is a special case where the moleculeID is 0
>The molecule ID is a second identifier attached to an atom.
>Normally, it is a number from 1 to N, identifying which molecule the atom belongs to.
>It can be 0 if it is a non-bonded atom or if you don't care to keep track of molecule assignments.

In the lammpswriter case, the resid was always starting at 0, which was
treating the first molecule "residue" as a non-molecule.

Now, an additional flag `moleculeID_offset` will let the user change
this starting offset value if they want to use the previous starting
value of 0. The default behavior now is to start all IDs at 1.