Use f-strings when applicable, add experimental
Previously, the breaking_change decorator was not properly formatted to
use f-strings.

It was setup to use f-string functionality, but the `f` was not present
out front of the string. This PR fixes this issue, and also adds an
experimenal_feature decorator.

This decorator is meant to flag certain methods that might change more
often/less often than a standard deprecation cycle. These methods should
be used at the users own risk and a warning is raised to inform them of
this.