Use `functools.wraps` to preserve docstring
With the addition of the `@experimental_function` wrapper to decorate
functions that are experimental and subject to frequent changes, the
docstring of the decorated function was not preserved.

This prevents the user from being able to see the `__doc__` and proper
`__name__` attributes of these methods. This PR includes the
`functools.wraps` method to ensure that the experimental functions are
properly documented in our docs and docstrings.