Generate bonds with freud (#969)
* Include freud as a way to generate bonds between two species

* Use f-strings when applicable, add experimental

Previously, the breaking_change decorator was not properly formatted to
use f-strings.

It was setup to use f-string functionality, but the `f` was not present
out front of the string. This PR fixes this issue, and also adds an
experimenal_feature decorator.

This decorator is meant to flag certain methods that might change more
often/less often than a standard deprecation cycle. These methods should
be used at the users own risk and a warning is raised to inform them of
this.

* Update environment yml files to include freud

* add tests

* add freud to the documentation env.yml

* Update mock modules

* Include freud as a way to generate bonds between two species

* Update environment yml files to include freud

* add tests

* add freud to the documentation env.yml

* Update mock modules

* include A-B filtering for particle names

* Include support for particle names, update tests

* Use `functools.wraps` to preserve docstring

With the addition of the `@experimental_function` wrapper to decorate
functions that are experimental and subject to frequent changes, the
docstring of the decorated function was not preserved.

This prevents the user from being able to see the `__doc__` and proper
`__name__` attributes of these methods. This PR includes the
`functools.wraps` method to ensure that the experimental functions are
properly documented in our docs and docstrings.

* include freud only as a dev dependency

* Include optional tests for freud.

* Implement fixes and suggestions from Bradley

* Change periodicity support, modify tests using extended bounding boxes.

* Include unit information for distance in both generate bond methods.

* Add notes section for experimental feature

Co-authored-by: Co Quach <43968221+daico007@users.noreply.github.com>