Allow user to pad bounding box for small cmpds
Raised by @daico007 in #1006, for very small molecules, the bounding
boxes will be so close to zero that the determinant of the box vectors
will be close to zero, signaling a co-linear box (psuedo-{1D, 2D}) and
raising an exception.

This PR provides an additional parmeter for the
`mb.Compound.get_bounding_box` method, `pad_box` which is a boolean to
avoid these co-linear systems by adding a 1nm pad to all box edges.

`pad_box` is `False` by default, and unit tests have been provided to
check both cases.

Closes #1006