Top down approach for identifying molecules
This PR adds two methods to a compound,
`_recursive_id_molecules`
and
`group_by_molecules`.
These functions let a user reconfigure an mBuild
compound using the top down approach of iterating through
children. The is_independent() method is checked on each children
until False is returned, which identifies when the hierarchy is made
of a bonded together "molecule". This will be of use when converting to
GMSO `Topology`, the subtop information can be parsed from this
reconfigured mBuild compound in order to speed up atomtyping.

These methods use the compound.name attribute to group molecules
together. With a generic name like "Compound", the top level child
will be a single compound.

**Note that these tests won't pass until PR #1037 is merged

* Add mbuild.compound.Compound._recursive_id_molecules method
* Add mbuild.compound.Compound.group_by_molecules method